requirements.txt
setuptools_rust
cryptography
python-binance
python-dotenv
python-coinmarketcap
